Move the functions for the routes to his own file
